requirements.txt
fastapi==0.116.1
langchain
langchain-openai
matplotlib
mlflow
numpy
Pillow
pydantic
requests==2.32.3
scikit-learn
sentence-transformers
streamlit==1.49.0
torch
torchvision
uvicorn==0.30.1